refactor(TableGroup, OrderTable): 엔티티 변환
