+++ README.md
@@ -78,8 +78,8 @@
   - [X] 메뉴 상품 dto 모델 추가
 - [ ] 테이블 관리 리팩토링
   - [ ] 모델
-    - [ ] TableGroup 엔티티 변환
-    - [ ] OrderTable 엔티티 변환
+    - [X] TableGroup 엔티티 변환
+    - [X] OrderTable 엔티티 변환
   - [ ] TableGroupRepository 사용하도록 수정
   - [ ] OrderTableRepository 사용하도록 수정
   - [ ] dto 모델 추가
