+++ src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -1,23 +1,40 @@
-package kitchenpos.domain;
+package kitchenpos.table.domain;
 
+import javax.persistence.Entity;
+import javax.persistence.ForeignKey;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+
+@Entity
 public class OrderTable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long tableGroupId;
+
+    @ManyToOne
+    @JoinColumn(name = "table_group_id", foreignKey = @ForeignKey(name = "fk_table_group_of_order_table"))
+    private TableGroup tableGroup;
+
     private int numberOfGuests;
+
     private boolean empty;
 
     public OrderTable() {
     }
 
     private OrderTable(Long tableGroupId, int numberOfGuests, boolean empty) {
-        this.tableGroupId = tableGroupId;
+        this.tableGroup = TableGroup.of(tableGroupId);
         this.numberOfGuests = numberOfGuests;
         this.empty = empty;
     }
 
     public OrderTable(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
         this.id = id;
-        this.tableGroupId = tableGroupId;
+        this.tableGroup = TableGroup.of(tableGroupId);
         this.numberOfGuests = numberOfGuests;
         this.empty = empty;
     }
@@ -39,11 +56,11 @@ public class OrderTable {
     }
 
     public Long getTableGroupId() {
-        return tableGroupId;
+        return tableGroup.getId();
     }
 
     public void setTableGroupId(final Long tableGroupId) {
-        this.tableGroupId = tableGroupId;
+        this.tableGroup = TableGroup.of(tableGroupId);
     }
 
     public int getNumberOfGuests() {
